[PC] Fix PETSC factorization to work with MPI
This change addresses a failure in PETSC library when running
SubCommAlgebraTest with MPI.
It turns out that the default solver package (MATSOLVERPETSC)
does not support full LU factorization, as a preconditioner,
when running in distributed way with MPI. As a workaround,
allow the Factorization to fallback to the additive Schwarz
method (PCASM), and use incomplete factorization (ILU).

This commit also contains changes in TestRunner/RunParallelTest
to align the verbosity level of test execution with the verbosity
setting of utopia library.

P.S. The acronym PC in the commit header stands for product care.