[bug] Include comm in Tpetra copy constructor
The behavior of vector/matrix copy constructor is different
in Petsc and Tpetra backends.
Petsc copies the communicator, while Tpetra does not. This result
in different behavior when the vector/matrix is used in subComm
groups. Petsc inherits the subComm, Tpetra does not.
The fix is to copy the communicator also in Tpetra copy constructor.